[INTERNAL][FIX] sap.ui.fl: Keep new variant author across soft reloads
The author name of a new variant is initially not available and is first
returned by the backend when the change is persisted.
In this case, the runtime instance is directly updated, however the
persisted storage response receives the outdated object without the
support user.
Make sure to update the storage response with a change that was enhanced
with the new information from the backend, so that it survives soft
reload scenarios where the flex objects are recreated from the storage
response.

SNOW: DINC0135482
CR-Id: 002075125900002598082024
Change-Id: I8f69efe60292ff8869806071a533eafe3d054149