[FIX] sap/ui/model/DataState.js: consider also the control dirty state of the parent data state
If a binding is used as a part of a CompositeBinding, it may happen
that the control data state of the composite binding is dirty, but
the control data states of all inner bindings are not dirty.
The OData V4 model is not aware of composite bindings. It checks
isControlDirty at the bindings that have been created by the model to
reset the invalid data state after e.g. resetting all changes.
Enable DataState#isControlDirty to consider also the control dirty state
of the parent data state (CompositeDataState).

http://localhost:8080/test-resources/sap/ui/core/demokit/sample/common/index.html?component=odata.v4.Products
http://localhost:8080/test-resources/sap/ui/core/internal/samples/odata/v2/Products/index.html?sap-language=en
Change e.g. the Weight to "foo" -> the input field is marked as invalid: "Enter a valid unit amount"
Click on "Reset Changes" -> Value is reset to the original value and invalid marker is removed

PS1: POC
PS2: TDD
PS3-4: Review comments

Change-Id: I3b211f4d50b20657a3d03aa16327697706fb1461
SNOW: CS20240008267927