[INTERNAL]sap.ui.fl/layout: CViz improvements for hide/reveal
The change indicators for hide/reveal were divided between the control
being hidden and its parent, leading to a confusing user experience.

Now the change indicators are shown on the control if it is visible, and
on the parent if it is hidden - so the user can see the entire list in
one place.

This change adjusts the generic hide/unhide change handlers and the
specific ones for the SimpleForm.

JIRA: TEAM42-367
Change-Id: I53b3eca90ca5c546e17ed3063ab501a0269481cb