[FEATURE]sap.m.ComboBox: Improve arrow up and down behavior
Change how items are selected on arrow down interaction in sap.m.Combobox and when the autocomplete property in sap.m.Input and sap.m.MultiInput is set to 'true'.
When the user presses the arrow down button now the next item is selected instead the first one.

Change-Id: I4f84eb7c0026bf69b6c45ce81edaa4334fc12d37
JIRA: BGSOFUIRILA-3932