[FIX] sap.f.SidePanel: fixed side content closing on mobile device
Problem: Side panel with one action item in mobile would be initially set to it's opened state and could not be closed. It was related to
scrolling and initial set of action bar expanded state, which is irrelevant when there is only one action item.

Solution: now the issue is fixed.

Test: Manual; unfortunately failing scenario cannot be tested with
qunit or visual tests.

BCP: DINC0193800
Change-Id: Ia87b2e2e50178a48752d0b18554e320f916019c0