[FIX] sap.m.SinglePlanningCalendar: navigate to correct month in islamic calendar
Issue: Previous/Next month navigation would navigate to the wrong month
because it would set the month of the islamic date to the gregorian
month we are expecting to navigate to.

Solution: Get the current month from the CalendarDate that is set to the
correct calendar type, instead of the gregorian oStartDate.

SNOW: DINC0256701
Change-Id: Ia83e4838db69bbbe2d8feeec789a51d5cad58204