[INTERNAL]sap.ui.fl: Delete changes related to control variants from backend
When a user-layer or draft control variant is removed, the variant and
all its related flex objects should be deleted in the backend. This can
be done since these variants do not have any dependency on them.

This is only possible in backends where condensing is enabled.

Since no backend supports the deletion of Variant FlexObjects yet,
this change still keeps the Variant and corresponding "setVisible=false"
change in the backend. This will be improved in a follow-up change,
where the deletion of the Variant FlexObject will be supported for
condensing-enabled backends.

JIRA: TEAM42-376
Change-Id: I4a5718ca7a061cffb708417b0ee8920f35d1ee76