[INTERNAL] sap.f.FlexibleColumnLayout: react on layoutData update
Problem: FE may not be able to set all LayoutData values for different
layouts before the first rendering of the FlexibleColumnLayout. In this
case, we need to react on a later change of the properties of
FlexibleColumnLayoutDataForDesktop/Tablet, as usually updates on
LayoutDatas does not invalidate the parent control.

Solution: We listen for properties updates on
FlexibleColumnLayoutDataForDesktop/Tablet and fire an event which is
handled by the FlexibleColumnLayout. In this event there are parameter,
which help us indentify what property is actually changed.
FlexibleColumnLayout reacts only on updates of the current layout and
media (desktop or tablet). It calls _resizeColumns with parameter for
switching off the animations in this case for performance optimization
(otherwise the resizing animation tooks 200ms). We only do resize, no
invalidation and rerendering also for performance optimization.

However, changes in the properties of the
FlexibleColumnLayoutDataForDesktop/Tablet happen also when user drags
the separators for resize the columns of the FlexibleColumnLayout. At
this moment a columnsDistributionChange event is fired and
FE/applications need to update the relevant LayoutData and its property,
corresponding to the current layout, acordingly. In this case, we
prevent second resizing.

Additionally, this change also fixes an error, which happens when
FlexibleColumnLayout is placed in an UIComponent, as base LayoutData's
implementation of invalidate method searches for parent of the parent of
the LayoutData and calls _handleEvent on it, but UIComponent does not
have such implementation. In our invalidate method of the
FlexibleColumnLayoutData we now inform the FlexibleColumnLayout (direct
parent) for the invalidation.

JIRA: BGSOFUIPIRIN-6672
Change-Id: Ica6474c56e660d34dba095fc56c3c419673d76f3