[FIX] sap.f.SidePanel: fixed resizing when Side Panel is left-positioned
Issue: When the Side Panel is positioned in left, the resizing handle
was not visible (as it was rendered outside of the screen), and the
resizing functionality didn't work as well because of improper
calculations related to opposite placement.

Solution: Now the resizing handle is visible and calculations are
correct. As a result resizing now works in poth Side Panel positions.

SNOW: DINC0263487
Change-Id: I0af01bb88684f14c262e90300e9dcc620a6d8329