[FIX] sap/ui/model/odata/v2/ODataListBinding: fire dataReceived event for aborted requests
If an ODataListBinding looses its context, a pending request has to
be aborted to avoid that the data is inserted into a wrong place.
A "dataReceived" event has to be fired in this case to inform the list
control to remove the busy indicator.

PS1: Downport to 1.120
PS2-3: Resolve merge conflict

Change-Id: I2c565edb56d8b2aa7b1e5a18e401cd59e7ff8de2
SNOW: DINC0246364
CR-Id: 002075125900002964162024
(cherry picked from commit 7c9dcd0cf9e60e2c6288ec3397dfbecc6771627c)