[FIX] sap.m.TabContainer: sync customData between outer and internal controls
Issue: There is previous change that transfers TabContainerItem
customData objects from outer TabContainerItem control to internal
TabStripItem control, but after that the customData does not exist
in outer control anymore.

Solution: All methods of TabContainerItem and TabContainer, related
to customData are overriden and now they manage customData directly
to internal controls, but it is available to outer controls too.
For that, the internal controls are created in init method of outer
controls to assure that they exist at the moment when customData is
managed.

SNOW: DINC0242422
Change-Id: If31ba22afbcb623e64118a700e064d1ed0dae516