[FIX] sap.m.ObjectStatus: new CSS class for languages with higher line heights
Issue: There were cuttings of the text decorations on some languages
with higher characters (like Thai)

Solution: After several research and PoC tasks, there is a proposal
for fixing this issue (see related JIRA links below). The solution
is to set CSS line-height to 1.6, but there must be some adjustments
in each control. This change provides sapUiHigherText CSS class and
adjustments for sap.m.ObjectStatus control. The newly introduced
class can be applied on a control level, or when this is not possible
- on parent container/control or even on BODY level.

Snippix for test: #795261

BCP: 2370130730
JIRA: BGSOFUIBALKAN-7799
JIRA: FIORIDESIGN-11358
JIRA: BGSOFUIBALKAN-8571
Change-Id: I5f03ed6dbbc2aecd0c1dd260040fd9330276c8f2