[INTERNAL] sap.m.Popover: adjust visual test
There was an inconsistency with the alignment of a sap.m.Link used inside the test caused by HBox's RenderType.
This change is designated to be used for testing newly generated reference images in order to determine if that was the problem causing the unstability of the test.

SNOW: DINC0271319
Change-Id: I47d9e2079ffddf24572088921df70cb775778e61