[FIX] sap.m.TimePicker: fix cursor position when there is mask and no entry
Problem: When there is mask set to the TimePicker, and there is no
value entered, sometimes on focus of the input field, the cursor
doesn't go to the beginning of the input field (and mask) because
of unwanted mouse movement during the click of the button.

There was previous fix but as we found lately, it doesn't cover
all cases.

Solution: If there is mask set and there is nothing entered in the
TimePicker's input, the cursor is forced to move at the beginning
of the input field in onmouseup event handler.

Test snippet: #389598

SNOW: CS20240007797996
Change-Id: I1b9c7b2912ec99e7a8f4fce73482a94a411c6b47
CR-Id: 002075125900003149252024