[FIX] sap.f.DynamicPage: Footer handling improved
Problem:
Flickering of the header content upon collapsing the header.
This happens as a side effect of a workaround for handling
the footer in the case when the app has enabled the
"fitContent" property but the actual content does
not fit its container [and instead overflows]

Solution:
Disabling the "fitContent" property on application-side
stops the flickering.
As a permanent solution, however, we make the following
solution on DynamicPage control-side:
--Disable the effect of the "fitContent" property to prevent
the app enabling the "fitContent" property when the *content
inside does not fit* its container [and instead overflows],
which caused the footer to overlap the content
--remove the [now redundant] old workaround for handling the
footer in the above case and replace with proper allocation
of space for the footer.

Cherry-picked from the following 3 changes:
Iedabb78a9d2bf4f55d02a2132e66bee766ee607e
Idafd341cd48a210b860d845b49d55ec76369fe98
Iffe08b716c3a7066795994ebd5e891d4e062d15e

BCP: 2270187596
BCP: 2380041732
BCP: 2380073093
SNOW: CS20240007936740
Change-Id: I8fd27e41036c904d7cc686db91109286c3ab515d
CR-Id: 002075125900003215832024