[INTERNAL] jsdoc-template: remove duplicates after `processingComplete`
Processing of @augments, @mixin and @borrows tags might produce new
duplicates which are not yet visible during the `parseComplete` event.

Moving the cleanup to `processingComplete` fixes this.

The merge of the additional data that the UI5 AST visitor collected,
is still done in `parseComplete` so that it is available when doclets
are cloned during the processing of @augments, @mixin and @borrows tags.

Additionally, one meaningless `@override` comment has been removed
as it now would replace the (more meaningful) generated documentation
after a @borrows.

Change-Id: I6699513a2dd7d974f470a50ed0ff8656f535be44