[INTERNAL] sap.m.TableSelectDialog: Pressing already selected item closes dialog
Problem: When pressing an already selected item with rememberSelections:true, causes the dialog to stay open, because selectionChange event is not fired and it should not be fired, because there is no change of selection, but the dialog should close.

Solution: Delegate the event to the table and prevent selectionChange and prevent selcetionChange event to be fired when already selected item is selected again. Upon selection close the dialog.

JIRA: BGSOFUIRODOPI-2120
Change-Id: I9ccba3d260917f322019f5e76feba2f84c4324ff