[FIX]: sap.m.Label: when the label is with percent width it works
When the width of the Label is set to percent (for example 15%)
the inner div was with width inherit, which makes the width of
the text 15% of 15% and the text truncated (when there is enough
space not to truncate).
Now the width of the inner div is 100% and the text fills all available space.
Samples and visual test added.

CR-Id: 002075125900003655092024
SNOW: CS20240008289040
SNOW: DINC0281147
Change-Id: I5e55e6ec7905d79f8ca3a5094190163e00f0a790
(cherry picked from commit bfdb81a1df81815a0d65e0d5f22a002dc8ad1cf6)