[INTERNAL] sap.ui.core.HyphenationPlayground: remove obsolete dependency
As a follow-up to change 6272089, this commit removes the previously missed dependency and import of "sap/base/util/UriParameters",
which was replaced with URLSearchParams in the previous commit.
The redundant import caused the visual tests to still fail, so this change is a follow-up as part of the UI5 2.0 tests adapations.

JIRA: BGSOFUIRODOPI-3316
Change-Id: Iadfcaec82137e7ca85133c42372a09ca9dc9e11c