[INTERNAL] odata.v4.TableDelegate: Support context-defining properties
Adds the attribute extension.additionalProperties to receive the context-defining properties, which will replace extension.customAggregate. extension.customAggregate is still supported for compatibility, but already removed from the documentation.

Additional properties must be groupable. An aggregatable-only property cannot be used, because it can't be requested (how to aggregate?). There would need to be a CustomAggregate. A context-defining CustomAggregate appears to be pointless. Therefore, these properties are only added to $$aggregation.group.

For a groupable property, these properties and the text property are added to "additionally" ($$aggregation.group["groupedProperty"].additionally).

For an aggregatable property, even if it's also groupable, these properties are added to $$aggregation.group, without any "additionally" properties. Because in fact the CustomAggregate is shown in the table, and not the groupable property with the same name.

Several restrictions apply and are documented in sap.ui.mdc.odata.v4.TablePropertyInfo.

Further changes:
- Additional properties of automatically added key properties are no longer added.

Change-Id: I1ff0ff378afab7fba03eaed16b8a79f98e466197
Jira: CPOUIFTEAMB-2428