[INTERNAL] sap.m.Label: adjust qUnit test
In version 1.71, the qUnit test was freezing due to the condition for hyphenated label not being met within the timeouts.
This issue was caused by the label being truncated instead of broken into paragraphs.
This commit adjusts the test as a workaround for the issue, in case we want to avoid merging change 6289487 in the 1.71 branch.

SNOW: DINC0289205
Change-Id: I6138ee5f5ab477516f8f309c4e0b38980ee81620
CR-Id: 002075125900003584022024