[FIX] Field with ValueHelp: number of items not announced
If the dropdown opens automatically when the Field is focused, the number of available items is not announced.

Now the number of items is annonced if a dropdown is opened and the exact number is known. (If there is a growing table the number of items is not known and cannot be announced in this case.)

Change-Id: I94b5097588ce3c16899b231fe683b88e56f5cd99
SNOW: DINC0277822