[FIX] sap.f.DynamicPageTitle: Fix focus lost on actions
Problem: When there is no enough space for all "actions" of
DynamicPageTitle to be shown and an overflow button is visible, if we
open the menu and click on a Button, opening a Dialog, and then close
the Dialog - focus gets lost.

Solution: handling the "focusfail" event on DynamicPageTitle level (as
it is fired from the Core to the parent of a Control, losing focus) and
redirect it to the OverflowToolbar (actual parent), containing the
"actions". OverflowToolbar already has implementation for handling the
"focusfail" event in such cases.
Redirecting the event is needed, as DynamicPageTitle overrides the
parent of its "actions".

SNOW: DINC0262573
Change-Id: I18a9a5e35c4527e687de37146f274f89c3ee1a15