[INTERNAL] sap.ui.documentation: Improve sample iframe url resolution
This simplifies the logic for resolving a custom iframe url for samples.
In addition, it solves previous limitations regarding the usage of query
parameters in the iframe url, which broke the RegExp causing an error.

Handling of relative path segments (e.g. "..") is now done by the
browser, so _resolveIframePath is no longer needed.

Change-Id: I6c81d106b3872d4f3ec228a87cd20db54d21dad3