[INTERNAL] sap.f.DynamicPageTitle: Safecheck added
Problem:
Error upon condensing 'move' actions when the 'content'
of the actions [to be condensed] does not contain the
expected 'targetAggregation' and 'targetContainer'
properties.

These properties are missing if the respective actions
were made with a UI5 version that did not (yet) support
condensing, because the above 2 properties were added
with change I68f1abb47a17121d35cccf18085ff16cf9e09ed9
that enables condensing.

Solution:
Since condensing is optional, fixed by omitting it
when the above 2 required properties are missing.
Tested that if the getCondenserInfo function
(that adds support for condensing) returns undefined,
then condensing is skipped.

SNOW: CS20240008438760
Change-Id: I8e85689fe220b60490a57b8ebd835e391e77debe