[FIX] sap.uxap.ObjectPageLayout: pin button is fixed
Problem: When header is pinned, then snapped and unsnapped, the pin icon
looks like the header is still pinned, while it is not (upon snapping it
actually gets unpinned).

Solution: Upon snapping we need to not only set the pressed state to the
pin button to false, but also to switch to the unpinned icon.
Additionally, when the provided snippix 770513 is runned on localhost,
the pin icon does not appear at all. That happens because the
themeChanged event is fired after the DynamicPageHeader is already
rendered and in this case the correct (pinned) icon is not set.
That's why now, when themeChanged event is fired, we set the correct
icon, depending on the pressed state.

SNOW: DINC0286334
Change-Id: I16f6898237bf2cb7f4a0ffc038e7e77b16b6f5af