[INTERNAL] OPA: Extend promiseWaiter by withResolvers support
The new withResolvers API directly returns our WrappedPromise. However, due to a condition that skips native Promise definitions, the auto waiter currently treats this promise as ignored. To address this, we've added an additional wrapper around the withResolvers API to ensure that the generated promise is correctly tracked.

Change-Id: Id8d84749647ea04d0393b3729a909c2c14593d8e
JIRA: BGSOFUIPIRIN-6670