[FIX] Demo Kit: Handle failed file fetches and display warning message
Problem:
When some files failed to fetch during the download process, the overall download was stopped, and no files were downloaded. This was causing incomplete downloads and preventing the user from retrieving the successfully fetched files, also some apps do not have a configuration property, yet the download link was still displayed

Solution:
Now, when files fail to fetch, the process continues by resolving the overall promise, allowing the successfully fetched files to be downloaded. A warning message box is displayed listing the files that could not be fetched, but the remaining files are still downloaded.
As for the download link, it is now conditionally visible based on the presence of the config property. If the config exists, the link will be displayed, otherwise, it remains hidden.

SNOW: DINC0271306
Change-Id: Idceca8bc13279bc0c7df758b6826a3da16f7037e