[INTERNAL][FIX]sap.ui.fl: Persist irrelevant properties in change object
The properties copied from the "content" change structure were being set
directly on the FlexObject (due to using the same references) -
which also caused the data to be persisted.

Also the error thrown when there is a property with the same name was
changed to a warning since this situation happens in existing change
handlers and this code will be removed in the future.

SNOW: DINC0304926
Change-Id: Ia03f459be02be4cf723bffb744ce83ec626a5b43