[FIX] sap.m.SinglePlanningCalendar: fixed week days position in 320px
Problem: There was a misplacement of week day names in the header of
the Month grid of SinglePlanningCalendar control because of fixed
margin-left of the container that contains week day names.

Solution: Now this is fixed by setting percentage widths for week day
names on smaller display widths.

SNOW: DINC0172416
Change-Id: I26f9dd0c269abf06c2cc447107515baaed904716