[FIX] sap.uxap.ObjectPageLayout: Fix scrolling when switching tabs
Problem: When expanded header is too big and user is switching tabs in
IconTabBar mode, ObjectPage is getting scrolled to the top, making the
IconTabHeader and all its content not visible, which is unpleasant from
user perspective.

Solution: When tab is clicked, we preserve the current scroll position
and then restore it if header is expanded amd the target section is the
first visible section (by default, previously, in this case, the
ObjectPage was scrolled to top).

SNOW: DINC0300739
Change-Id: I73d748dccc14a6d5d8af51803a8967ba0e46372b