[INTERNAL] Table: Fix unstable v4 selection tests
On displays with a refresh rate > 60hz, the UI can be ready before the binding (requestAnimationFrame is used in test utils).
Because of that, there might be no expandable rows yet when the test want's to expand them.

To fix the instability, the test waits for a binding change explicitly, before waiting for the table to update after that change.

Change-Id: I0b248a885806e619b9ab860fd7a1867979b9cc45