[FIX] sap.m.Toolbar: Fixed active toolbar focus jump
Problem: If Toolbar is active and out of the viewport, if the focus
land on it, the page is scrolled to its very top.

Root Cause: The button for the focus reference is outside of the viewport.

Solution: We make the the focus reference button to be positioned inside
the Toolbar.

SNOW: DINC0237720
Change-Id: I72e45e19e07c6a54e2c96c9db5bb88bc78b810de