[FIX] sap.m.ViewSettingsDialog: fixed focus after pressing the Reset button
Issue: When pressing the Reset button of the ViewSettingsDialog, in
some cases (when there are no custom items or tabs) it goes disabled,
and the focus goes on the first focusable element. But when there are
custom items or tabs, the Reset button stays enabled, but the focus
is moved again to the next focusable element.

Solution: now there is a check for custom items/tabs, and if there
are, the focus is not moved from the Reset button.

SNOW: DINC0099607
Change-Id: I4c8664d563570a857a1b2e714d25aa9cd8d34020