[INTERNAL][FIX] CellSelector: Account for invisible content
If a cell had content with "visible=false", the necessary classes were not applied to the td element, as there was no way of determining the td element from the cell content due to it not being rendered/present in the DOM.

Therefore, if no DOMRef can be found, we need to have a look at the DOMRef of the row itself and search for the related td element.

Also need to account for fixed columns in GridTable. We need to search for the equivalent row DOM ref in the fixed area.

SNOW: DINC0294699
Change-Id: Ie261fd3b4b2c327146ccdc6ecfaba693895b60c8