[FIX] sap.m.SinglePlanningCalendarGrid: first week of year is now correct
Issue: the week numbering in en-US locale is different than in other
locales, and there might be a week at the end of the year/begin of
the next year that is numbered 53 (even 54) in current year and 1 in
the next year. That's why sometimes week numbers look inconsistent
when first mont of a week is diplayed - for example 53, 2, 3, 4, ...
The inconsistency is visible when you display a month in SPC Month
grid and open DatePicker in the calendar header.

Solution: now in the December of an year that have 53/54 week, it is
displayed as 53/54, but in the January week numbers start with 1.

SNOW: DINC0142780
Change-Id: I1f32901d5001dea0596eeeb812f9936f88743485