[FIX] sap.ui.unified: Fix editing of DemoKit Samples
Problem: DemoKit Samples of sap.ui.unified library cannot be edited with
the DemoKit's editing functionality.

Solution: All samples are moved in directory /demokit/sample/, because
the service worker of the DemoKit is working specifically and only with
samples, placed in the demokit folder.

All samples are tested and work correctly.

SNOW: DINC0299676
Change-Id: I75d8eaa790e1da592cea1f73e99123a462133d6d