[FIX] sap.m.Image: Inline svg rendering issue corrected
Problem:
Some elements of the inline SVG were not rendered.
This happened when the SVG contained relative links
to other resources. This issue occurred because
the relative links were automatically resolved
relative to the document’s URL instead of the
SVG image’s URL.

Solution:
Added a conversion process for the relative URLs
within the SVG source to absolute URLs, using the
SVG image’s href as the base URL for this conversion.

SNOW: DINC0309668
Change-Id: I23236fc18528691e082acbc1aede947627dd40ae