[FIX] sap.m.OverflowToolbar: Do not show empty Popover
Problem: WHen only ToolbarSeparators are moved in the overflow Popover,
they are not visible (hidden with CSS) and the Popover menu looks empty.

Solution: First child and last child ToolbarSeparators in the Popover
menu are hidden with CSS, as they are actually not needed (we only need
to visually separate different controls from one another). That's why,
in case there are only ToolbarSeparators placed in the Popover, we do
not show the overflow menu button, as the Popover looks empty otherwise.

SNOW: DINC0272039
Change-Id: Ib7c90f4ef0a60b083c8ccf8c5ea2789cf09593bd