[INTERNAL] core/Element.qunit.js: Introduce Timeout for FF
Test scenario:
When the last button is being disabled, no new focus target
can be found via the "onfocusfail" handler.
The focus is expected to be shifted to the document body,
triggered via the finalization of the rendering and
FocusHandler#restoreFocus.

In FF the focus shift to the body seem to happen async.
=> Introduced a setTimeout to wait for the focus movement
to happen.

SNOW: DINC0331513
Change-Id: I58de2880eaf8109c098855ed23b158cc87329123