[FIX] sap.uxap.ObjectPageLayout: Incorrect scrolling prevented
Problem:
Undesired scrolling to the bottom of the ObjectPage
after hiding the footer-toolbar of the page in the
case when the focus was inside the footer-toolbar
prior to hiding the toolbar.
Hiding the footer-toolbar let to loosing the focus
(as the document.activeElement was inside the same
toolbar). This, by design, triggered the 'onfocusfail'
event handler of the footer's parent (namely, the
objectPage) that should move the lost focus to the
next focusable neighbour of the toolbar. However,
the default implementation of 'onfocusfail' handler
does not account for the fact that the toolbar is
*positioned absolutely* on top of scrollable content
=> it ends up moving the focus to an element of the
scrollable content that is close to the footer
(in terms of DOM order) BUT outside the scrollport
=> eventually leads to undesired scrolling to
the newly focused element.

Solution:
A final solution can be to provide an alternative
implementation of 'onfocusfail' that restores the
focus to a visible element inside the scrollport
(that may be a subject of a BLI and may also be
implemented on a more general level than ObjectPage
as the same issue can occur in other controls).
Since the issue in the ticket is preventing the
undesired scrolling, fixed by omiting the default
logic in the above case to prevent the scroll.
Restoring the focus is not an issue in this case
for now as the app itself handles the focus.

SNOW: DINC0264581
Change-Id: I984a8069d5ac17a6033ebe1793068d13033540a9
(cherry picked from commit 153b96be844e67b76514086283f77f3fba412f1e)
CR-Id: 002075125800003971972024