[INTERNAL][FIX] sap.ui.fl: Fix timing issue with FlexState initialization
When initializing the FlexState, make sure that creating a new state
based on the response and the full initialization (initializeNewInstance
filling storageResponse and runtimePersistence) happen synchronously.
Otherwise it can happen that the incomplete state is accessed by a
DataSelector in the meantime due to unfortunate timing.
Realistically this can happen during the start up hooks, where multiple
inits are performed with async callbacks that get scheduled into the
same promise cycles as subsequent initalize calls.

SNOW: CS20240008590588
Change-Id: I6d984002bfd946ce48824f8fc649cac52adf83ab