[FIX] sap.uxap.ObjectPageLayout: Calc FIX
Problem: In certain situation, when the DOM element
of ObjectPageLayout is still not rendered/ rendered
in hidden parent element, onBefore rendering
we are storing the scrolling offset based on its
current scroll position (0) minus the stored
position of the targeted subsection, so we end up
with negative scroll position and wrong stored
section.

Solution: We double check for real DOM element
painted on the document.

SNOW: CS20240008278829
Change-Id: I534f67475c89b8ac316e0c58cc8937dc8612d2fd