[FIX] sap.uxap.ObjectPageSubSection: Focus loss prevented
Problem:
Focus loss upon re-renedring of the subSection,
if the focus was inside one of the subSection
blocks prior to the re-rendering.
This happends because upon each rerendering
the subsection temporarily removes all its blocks.
When the block holding the focus is [temporarily]
removed, the default onFocusFail handler moves the
focus away from the removed block and into to the
next focusable element (as the handler  does not
know that the block was removed only temporarily).

Solution:
Refactored to not use temporary removal of blocks.
Instead:
1) the blocks are now removed only when
removal requested by API call => enhanced the
_onRemoveBlock (and respectively the _onAddBlock)
functions to also remove/add blocks from/to the
actual control-tree (as otherwise they are only
buffered to the private 'aggregationProxy' array).
2) the content of the blocks is reset (temporarily
removed and then added again) only when resuming
the propagation of data-binding updates, as a
workaround way to catch up updates that were missed
when propagation was being skipped.

SNOW: DINC0335494
Change-Id: Iaa76a751ec2f6aec058ee02c1cad9a10fb702029