[FEATURE] We've now introduced the uploadProgress and uploadTerminated events for uploadItem within the UploadSetwithTable plugin
With this change, the plugin introduces uploadProgress and uploadTerminated event at the sap.m.UploadItem level. The events notifies the upload progress or the upload termination information of the file, based on the XHR upload event. The consumer of the plugin can attach event handlers to track the continous progress of the file upload and update their UI accordingly. Also, the consumers can attach uploadTerminated event to get notified when an upload is terminated. These changes allow consumers to have custom UIs for the progress bar and the terminate button. The uploadTermintated event is implemented in collaboration with the terminateUload API. Alternatively, consumers can also listen to the progress and terminate events available with the uploaderTableItem class for the same progress and termination information.

JIRA: FIORITECHE1-8700
Change-Id: I74f46aef46712230e9edb12fe87243995030c0b5