[INTERNAL][FIX]sap.ui.rta: Temporary removal of a changeVisualization test
There has been an issue with one of the tests in the changeVisualization test suite:
"when a change indicator with a related indicator (two display controls for the same change) is hovered/focused"
There is also a general issue in this Test:
The test is used to make sure that the focus is set correctly when the DetailsPopover gets opened or closed. In the
current implementation, the wrong popover gets opened and therefore is not valid. The test needs to be reworked to correctly open the DetailsPopover.
As the test is currently not working, it has been removed to allow the core colleagues to merge their changes to the popover handling.

Change-Id: I45f1bd3ccaf52bf2b261c85d791808eb2c0c684b