[INTERNAL][FIX] plugins.ODataV4Selection: Selection after enablement
If the the plugin is disabled while there is a pending selectionChange
timeout, the timeout is cleared to not fire a selectionChange event. If
the plugin is then enabled, selectionChange events are not fired
anymore, because the timeout is not cleaned up correclty on
deactivation.

The issue occured in the MDC Table, where FE set the selectionMode to
"None" while a selection change was pending. Afterwards FE enabled
selection again.

SNOW: DINC0338964

Change-Id: I02ef830a637d09de40b905761c65ba446b752a88