[INTERNAL] odata.v4.TableDelegate: Hide additionalProperties attribute
The attribute was introduced for the id+text scenario, where the text depends on the id (via context-defining properties annotation). The exact use-cases and how to handle them are not yet clear enough, therefore the API is subject to change and the attribute is hidden.
FE can still use it as before.

Change-Id: I6ec535d4689f73e95373a1d86458a3c36bd7bf7c