[FIX] sap.m.DateTimePicker: YearPicker content now fits in limited size
Issue: There was a fixed width in rem set to the Year/YearRangePicker
container and when there is zoom (>100%), the years/year ranges that
are inside of the container didn't fit and were cut.

Solution: Now the width of the container is set to "auto" and the
content fits without a cut.

SNOW: DINC0343409
Change-Id: I200ee49abfa0a0e2e6b4a0f31fd830c382b5db5c