[INTERNAL] sap.m.IllustratedMessage: Improved metadata config
The metadata.json file associated with illustration sets is much more
flexible now.

There can be custom configuration for any symbol for a specific
path/theme.

The configuration might consist of symbol and/or size substitution.

Additionally, a metadata.json file is created for the default "sapIllus"
illustration set.

Previously the default set was hard coded in the IllustrationPool, now
it is required externally via the IllustrationPool _registerDefaultSet
helper function.

JIRA: BGSOFUIPIRIN-6726

Change-Id: I7ba95b227c0882f649ce245abd9274fdf4f9b55d