[INTERNAL] We've now optimized file preview performance with the UploadSetwithTable plugin
With this change the file preview dialog will load only the content of current page in focus and only loads the content of the page when carousel is moved to that page which optimises the perfomance of file preview dialog loading. Each control is wrapped under VBOX container and the content of the VBox is loaded with actual control on focus of that page. Change includes refactoring of the code to move the specific control loading to reusable method & introduing VBox container wrapper for each page control, the in memory tracking of loaded pages so the same page is not loaded again till the file preview dialog instance is alive.

JIRA: FIORITECHP1-31984
Change-Id: I50d81a683822bac138c5ad5ef0f94c1827c0ee28