[INTERNAL] sap.ui.fl: remove dependency of VariantModel to the write package
The VariantModel is part of the initial fl library bundle and, together with all
of the dependencies, is loaded on app start. The write package is only needed
on user interactions, so the VariantModel should not have a direct dependency to
the write package.
The logic for the write package is now moved to the VariantManager, which
uses the VariantModel directly.

This commit does not completely move all the write related logic,
only the logic that relies on the write package.

JIRA: TEAM42-336
Change-Id: I9c79c1a8546439055fdb2c602cb40561daae83be