[INTERNAL][FIX] sap.ui.rta: Fix editableField size comparison in Safari
When the RenameHandler calculates the size of the editableField, it
usually uses the offsetWidth of the associated domRef of the renamed
element and only if it is larger than the overlay, falls back to the
parent instead.
For the VariantManagement this is not the case, however the test was
still asserting against the parent.
In this scenario, the parent has a different size on Safari (off by one
pixel - probably due to different box-sizing), thus the wrong value was
compared.
Furthermore, the test now properly asserts the offsetWidth (including
e.g. padding) of the editableField instead of the boundingClientRect
width, since the RenameHandler also uses the offsetWidth.

SNOW: DINC0348023
CR-Id: 002075125800004330372024
Change-Id: I8a1abf9d0ea72dfc16100c2d111245ac2f67372e
(cherry picked from commit 89a4fc2911d9acad14bfc09a352a7e6b68d7c1c8)