[FIX] sap.m.Menu: Connection between sap.m.MenuItem and sap.ui.unified.MenuItem is handled correctly now
Problem: After removing items from the menu and later adding them, those were not visualised in it.

Solution: the problem came from the fact that the items weren't added in the unified menu items aggregation because the check that suppose to add them was checking if they only exist but not if they are presented in the aggregation.

SNOW: DINC0294130
Change-Id: Id995281f078d4aa7a6a4500f79b8f3f031ff27b3