[FIX] odata.v4.TableDelegate: Clear selection when updating the binding
When changes are made in the binding with multiple API calls, the binding fires a change event with the consolidated reason "change". The information that there is a sort or filter change is lost, hence the inner GridTable does not clear the selection. The changes could affect the indices and make the current selection invalid. Therefore, the delegate has to clear the selection.

(cherry picked from commit ad7f878990ddd5aef6500391d33ab0653c1657b1)

Change-Id: I897d49c0f34c8d7ad5066ccd9c3a873ee66ae2ca
SNOW: CS20250009612490
CR-Id: 002075125900001337382025