[INTERNAL] ListBase: remove selections during sort and filter
 - rememberSelections property of the list retains the selection of list
items after a binding update.
 - In UI5, ManagedObject#updateAggregation reuses existing templates and
only sets the binding context of controls. This means, if a property is
not bound (e.g. selected property of the ListItemBase) and updated
programmatically might stay untouched although the context of the
control is changed after the binding update.
 - If rememberSelections of the list is false and if the selected
property of list items is not bound two-way then filtering and sorting
on the binding should clear previous selection.
 - Documentation of rememberSelection is updated to indicate the new
behavior

SNOW: DINC0352752

Change-Id: I9a7be663b5dff5229c005a30d04ffd9966013c79