[INTERNAL][FIX] - sap.ui.fl: Implement correct behavior for all contexts provided
This change fixes several issues that could lead to data loss in context
of context-based adaptation (adapting uis for specific user roles) and
context based views (adapt views). When user creates new views (adapt
views) or context-based adaptation (adapting ui for specific user roles)
and clicks on a go button in the smart filter bar or switches to another
version, internally a reinitialisation might happen, that deletes all
changes done until this point (data loss). This is due to a not
correctly set property in the FlexState. This has been fixed with this
change.

Change-Id: I7628d2bec6099e4e8080e659b96efbdf50630c3b
SNOW: DINC0252294
CR-Id: 002075125800001188682025